refactor: 검증 도메인 로직 추가
- 도메인 역할에 맞는 패키지로 SeatValidator로 이동 및 클래스명 변경
- ScheduleValidator로 콘서트 예약 시간대 검증
