+++ src/main/java/io/project/concertbooking/domain/concert/SeatValidator.java
@@ -1,21 +1,20 @@
-package io.project.concertbooking.domain.reservation;
+package io.project.concertbooking.domain.concert;
 
 import io.project.concertbooking.common.exception.CustomException;
 import io.project.concertbooking.common.exception.ErrorCode;
-import io.project.concertbooking.domain.concert.Seat;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
 
 @Component
-public class ReservationValidator {
+public class SeatValidator {
 
-    public void checkIfSeatAvailable(List<Seat> seats) {
+    public void checkIfSeatsReservable(List<Seat> seats) {
         seats.stream()
                 .filter(seat -> seat.isReserved() || seat.isOccupied())
                 .findFirst()
                 .ifPresent((seat) -> {
-                    throw new CustomException(ErrorCode.SEAT_NOT_AVAILABLE);
+                    throw new CustomException(ErrorCode.SEAT_UNAVAILABLE);
                 });
     }
 }
