+++ src/test/java/io/project/concertbooking/domain/concert/ScheduleValidatorTest.java
@@ -1,0 +1,4 @@
+import static org.junit.jupiter.api.Assertions.*;
+class ScheduleValidatorTest {
+  
+}
